Fold `native-instance-utils.ts` into `native-conversion.ts`.
...and do various other renaming and tweakage.
