+++ packages/data-model/deno.json
@@ -29,7 +29,6 @@
     "./deep-freeze": "./src/deep-freeze.ts",
     "./EmptyReconstructionContext": "./src/EmptyReconstructionContext.ts",
     "./fabric-instances": "./src/fabric-instances/index.ts",
-    "./native-instance-utils": "./src/native-instance-utils.ts",
     "./fabric-primitives": "./src/fabric-primitives/index.ts",
     "./fabric-type-tags": "./src/fabric-type-tags.ts",
     "./fabric-value": "./src/fabric-value.ts",
