+++ packages/data-model/src/native-conversion.ts
@@ -7,11 +7,6 @@ import {
 } from "./interface.ts";
 import { isFabricValue } from "./type-check.ts";
 import { FabricEpochNsec } from "./fabric-primitives/FabricEpochNsec.ts";
-import {
-  errorClassFromType,
-  isConvertibleNativeInstance,
-  UNSAFE_KEYS,
-} from "./native-instance-utils.ts";
 import { FabricError } from "./fabric-instances/FabricError.ts";
 import { FabricNativeWrapper } from "./fabric-instances/FabricNativeWrapper.ts";
 import { FabricRegExp } from "./fabric-primitives/FabricRegExp.ts";
@@ -20,6 +15,7 @@ import { NATIVE_TAGS, tagFromNativeValue } from "./native-type-tags.ts";
 import { isArrayWithOnlyIndexProperties } from "@commonfabric/utils/arrays";
 import { cloneHelper } from "./value-clone.ts";
 import { isDeepFrozenFabricValue } from "./deep-freeze.ts";
+import { FrozenSet } from "./frozen-builtins.ts";
 
 /**
  * Helper for `shallowFabricFromNativeValue()`, which rejects native objects
@@ -31,6 +27,56 @@ function rejectExtraProperties(value: object, typeName: string): void {
       `Cannot store ${typeName} with extra enumerable properties`,
     );
   }
+}
+
+/**
+ * Returns `true` if the value is a native JS object type that the fabric
+ * system knows how to wrap. These are the "wild-west" instances that get
+ * converted into `FabricNativeWrapper` subclasses, `FabricPrimitive` types,
+ * or `FabricInstance` types by the conversion layer.
+ *
+ * Arrays, plain objects, objects with `toJSON()`, and system-defined special
+ * primitives are recognized by `tagFromNativeValue()` but are NOT convertible
+ * native instances -- they have their own handling paths in the conversion
+ * layer.
+ */
+export function isConvertibleNativeInstance(value: object): boolean {
+  switch (tagFromNativeValue(value)) {
+    case NATIVE_TAGS.Error:
+    case NATIVE_TAGS.Map:
+    case NATIVE_TAGS.Set:
+    case NATIVE_TAGS.Date:
+    case NATIVE_TAGS.Uint8Array:
+    case NATIVE_TAGS.RegExp:
+      return true;
+    default:
+      return false;
+  }
+}
+
+/** Keys that must never be copied to prevent prototype pollution. */
+export const UNSAFE_KEYS: FrozenSet<string> = new FrozenSet([
+  "__proto__",
+  "constructor",
+]);
+
+/** Map from Error subclass name to its constructor. */
+const ERROR_CLASS_BY_TYPE: ReadonlyMap<string, ErrorConstructor> = new Map([
+  ["TypeError", TypeError],
+  ["RangeError", RangeError],
+  ["SyntaxError", SyntaxError],
+  ["ReferenceError", ReferenceError],
+  ["URIError", URIError],
+  ["EvalError", EvalError],
+]);
+
+/**
+ * Helper for `FabricError.[RECONSTRUCT]()`, which returns the `Error`
+ * constructor for the given type string (e.g. `"TypeError"`). Falls back
+ * to the base `Error` constructor for unknown types.
+ */
+export function errorClassFromType(type: string): ErrorConstructor {
+  return ERROR_CLASS_BY_TYPE.get(type) ?? Error;
 }
 
 /**
